Made changes to port the program to windows.

+++ bits/SpaceGroup/Exceptions.cpp
@@ -62,7 +62,7 @@ MyException::~MyException() throw()
 
 void MyException::addError(char* pErrMsg)
 {
-    int tLen = strlen(iErrStr)+strlen(pErrMsg)+2;
+    int tLen = (int)(strlen(iErrStr)+strlen(pErrMsg))+2;
     char* tNewStr = new char[tLen];
     sprintf(tNewStr, "%s\n%s", pErrMsg, iErrStr);
     delete iErrStr;
